Refactor: Consolidate duplicated logic for dayToken and exercise extraction
