+++ .jules/hunter-progress.md
@@ -1,0 +1,6 @@
+# Hunter Progress
+
+## Session 1
+- **Fixed:** Duplicated `dayToken` logic in `src/utils/dayToken.ts` and `scripts/day-token.js`. Consolidated into `src/utils/dayToken-core.js`.
+- **Fixed:** Duplicated exercise extraction logic in `src/utils/contentLoader.ts` and `scripts/validate-content.js`. Consolidated into `src/utils/exercise-extractor-core.js`.
+- **Verified:** Build, lint, and tests pass.
