+++ src/utils/dayToken-core.d.ts
@@ -1,0 +1,7 @@
+export function normalizeDayToken(value: any): string
+export function parseDayToken(value: any): { token: string; number: number; suffix: string; sortKey: string } | null
+export function compareDayTokens(a: any, b: any): number
+export function dayTokenFromPath(path: any): string | null
+export function extractDayToken(value: any): string | null
+export function dayTokenFromReference(value: any): string | null
+export function dayTokenToProgressId(value: any): number
