+++ src/utils/exercise-extractor-core.d.ts
@@ -1,0 +1,9 @@
+export interface ExtractedExercise {
+  title: string
+  goal: string
+  starterCode: string
+  expectedOutput?: string
+  tags: string[]
+}
+
+export function extractExercisesFromContent(content: string): ExtractedExercise[]
